% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{.assert_predictors}
\alias{.assert_predictors}
\title{This assertion checks that the \code{.predictors} columns
are all contained within \code{newdata}}
\usage{
.assert_predictors(newdata)
}
\description{
This assertion checks that the \code{.predictors} columns
are all contained within \code{newdata}
}
\keyword{internal}
