% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{.assert_valid}
\alias{.assert_valid}
\title{This assertion checks that the columns of \code{newdata}
do not contain any of: NaNs, NAs, Infinities}
\usage{
.assert_valid(newdata)
}
\description{
This assertion checks that the columns of \code{newdata}
do not contain any of: NaNs, NAs, Infinities
}
\keyword{internal}
