% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{.is_assert_disabled}
\alias{.is_assert_disabled}
\title{Checks if assertions are disabled. If \code{disable_assert}, or \code{..env}
cannot be found, then it defaults to FALSE.}
\usage{
.is_assert_disabled()
}
\description{
Checks if assertions are disabled. If \code{disable_assert}, or \code{..env}
cannot be found, then it defaults to FALSE.
}
\keyword{internal}
