% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{.model_exists}
\alias{.model_exists}
\title{Check if the model exists in the cache directory}
\usage{
.model_exists()
}
\description{
Check if the model exists in the cache directory
}
\keyword{internal}
