% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\docType{data}
\name{.predictors}
\alias{.predictors}
\title{Expected predictors}
\format{
An object of class \code{character} of length 5.
}
\usage{
.predictors
}
\description{
Expected predictors
}
\keyword{internal}
