% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{download_model}
\alias{download_model}
\title{Download the GBM roughness model}
\usage{
download_model()
}
\description{
If the model does not exist in the cache directory,
this function invokes the internal download model function
to pull it from a CDN. If the model does exist, then this
function performs no operation.
}
