% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{hr_augment}
\alias{hr_augment}
\title{Augment data with GBM-based roughness predictions.}
\usage{
hr_augment(newdata, ..., override_path = NULL)
}
\arguments{
\item{newdata}{(\code{data.frame})\cr
Must contain the correct predictor columns, i.e.
"pathlength", "arbolatesu", "lengthkm", "areasqkm", "slope"}
}
\value{
\code{newdata} with \verb{$.pred} column containing the predictions.
}
\description{
Augment data with GBM-based roughness predictions.
}
