% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{hr_predict}
\alias{hr_predict}
\title{Predict roughness values using GBM model.}
\usage{
hr_predict(newdata)
}
\arguments{
\item{newdata}{(\code{data.frame})\cr
Must contain the correct predictor columns, i.e.
"pathlength", "arbolatesu", "lengthkm", "areasqkm", "slope"}
}
\value{
\code{data.frame} with the column \verb{$.pred} containing the predictions.
}
\description{
Predict roughness values using GBM model.
}
